// src/propietarios/propietarioMascotaResponse.dto.ts
export class PropietarioMascotaResponseDto {
  idMascota: number;
  nombre: string;
  especie: string;
  raza: string;
  edad: number;
  genero: string;
  peso: number;
  color: string;
  fechaNacimiento: Date;
  microchip: string;
  estado: string;
  notas: string;
  fechaUltimaVisita: Date;
  foto: string;
}